backend/requirements.txt
Flask
requests
responses
flask-cors
python-dotenv